refactor(sql): move dialects to always-importable location
