+++ ibis/backends/duckdb/__init__.py
@@ -27,7 +27,7 @@ from ibis.backends.base import CanCreateSchema, UrlFromPath
 from ibis.backends.base.sqlglot import SQLGlotBackend
 from ibis.backends.base.sqlglot.compiler import STAR, C, F
 from ibis.backends.duckdb.compiler import DuckDBCompiler
-from ibis.backends.duckdb.datatypes import DuckDBPandasData
+from ibis.backends.duckdb.converter import DuckDBPandasData
 from ibis.expr.operations.udf import InputType
 
 if TYPE_CHECKING:
