+++ ibis/backends/impala/tests/snapshots/test_value_exprs/test_timestamp_deltas/days/out2.sql
@@ -1,3 +1,3 @@
 SELECT
-  `t0`.`i` - INTERVAL '5' DAY AS `TimestampSub(i, 5D)`
+  `t0`.`i` - INTERVAL 5 DAY AS `TimestampSub(i, 5D)`
 FROM `alltypes` AS `t0`
