+++ ibis/backends/impala/tests/snapshots/test_value_exprs/test_timestamp_deltas/hours/out2.sql
@@ -1,3 +1,3 @@
 SELECT
-  `t0`.`i` - INTERVAL '5' HOUR AS `TimestampSub(i, 5h)`
+  `t0`.`i` - INTERVAL 5 HOUR AS `TimestampSub(i, 5h)`
 FROM `alltypes` AS `t0`
