+++ ibis/backends/impala/tests/snapshots/test_value_exprs/test_timestamp_deltas/minutes/out2.sql
@@ -1,3 +1,3 @@
 SELECT
-  `t0`.`i` - INTERVAL '5' MINUTE AS `TimestampSub(i, 5m)`
+  `t0`.`i` - INTERVAL 5 MINUTE AS `TimestampSub(i, 5m)`
 FROM `alltypes` AS `t0`
