+++ ibis/backends/impala/tests/snapshots/test_value_exprs/test_timestamp_deltas/months/out2.sql
@@ -1,3 +1,3 @@
 SELECT
-  `t0`.`i` - INTERVAL '5' MONTH AS `TimestampSub(i, 5M)`
+  `t0`.`i` - INTERVAL 5 MONTH AS `TimestampSub(i, 5M)`
 FROM `alltypes` AS `t0`
