+++ ibis/backends/impala/tests/snapshots/test_value_exprs/test_timestamp_deltas/seconds/out2.sql
@@ -1,3 +1,3 @@
 SELECT
-  `t0`.`i` - INTERVAL '5' SECOND AS `TimestampSub(i, 5s)`
+  `t0`.`i` - INTERVAL 5 SECOND AS `TimestampSub(i, 5s)`
 FROM `alltypes` AS `t0`
