+++ ibis/backends/impala/tests/snapshots/test_value_exprs/test_timestamp_deltas/weeks/out2.sql
@@ -1,3 +1,3 @@
 SELECT
-  `t0`.`i` - INTERVAL '5' WEEK AS `TimestampSub(i, 5W)`
+  `t0`.`i` - INTERVAL 5 WEEK AS `TimestampSub(i, 5W)`
 FROM `alltypes` AS `t0`
