+++ ibis/backends/impala/tests/snapshots/test_value_exprs/test_timestamp_deltas/years/out2.sql
@@ -1,3 +1,3 @@
 SELECT
-  `t0`.`i` - INTERVAL '5' YEAR AS `TimestampSub(i, 5Y)`
+  `t0`.`i` - INTERVAL 5 YEAR AS `TimestampSub(i, 5Y)`
 FROM `alltypes` AS `t0`
