+++ ibis/backends/oracle/compiler.py
@@ -6,13 +6,12 @@ import sqlglot as sg
 import sqlglot.expressions as sge
 import toolz
 from public import public
-from sqlglot.dialects import Oracle
-from sqlglot.dialects.dialect import create_with_partitions_sql, rename_func
 
 import ibis.common.exceptions as com
 import ibis.expr.operations as ops
 from ibis.backends.base.sqlglot.compiler import NULL, STAR, SQLGlotCompiler
 from ibis.backends.base.sqlglot.datatypes import OracleType
+from ibis.backends.base.sqlglot.dialects import Oracle
 from ibis.backends.base.sqlglot.rewrites import (
     Window,
     exclude_unsupported_window_frame_from_ops,
@@ -26,51 +25,11 @@ from ibis.backends.base.sqlglot.rewrites import (
 )
 
 
-def _create_sql(self, expression: sge.Create) -> str:
-    # TODO: should we use CREATE PRIVATE instead?  That will set an implicit
-    # lower bound of Oracle 18c
-    properties = expression.args.get("properties")
-    temporary = any(
-        isinstance(prop, sge.TemporaryProperty)
-        for prop in (properties.expressions if properties else [])
-    )
-
-    kind = expression.args["kind"]
-    if (obj := kind.upper()) in ("TABLE", "VIEW") and temporary:
-        if expression.expression:
-            return f"CREATE GLOBAL TEMPORARY {obj} {self.sql(expression, 'this')} AS {self.sql(expression, 'expression')}"
-        else:
-            # TODO: why does autocommit not work here?  need to specify the ON COMMIT part...
-            return f"CREATE GLOBAL TEMPORARY {obj} {self.sql(expression, 'this')} ON COMMIT PRESERVE ROWS"
-
-    return create_with_partitions_sql(self, expression)
-
-
-Oracle.Generator.TRANSFORMS |= {
-    sge.LogicalOr: rename_func("max"),
-    sge.LogicalAnd: rename_func("min"),
-    sge.VariancePop: rename_func("var_pop"),
-    sge.Variance: rename_func("var_samp"),
-    sge.Stddev: rename_func("stddev_pop"),
-    sge.ApproxDistinct: rename_func("approx_count_distinct"),
-    sge.Create: _create_sql,
-    sge.Select: sg.transforms.preprocess([sg.transforms.eliminate_semi_and_anti_joins]),
-}
-
-# TODO: can delete this after bumping sqlglot version > 20.9.0
-Oracle.Generator.TYPE_MAPPING |= {
-    sge.DataType.Type.TIMETZ: "TIME",
-    sge.DataType.Type.TIMESTAMPTZ: "TIMESTAMP",
-}
-Oracle.Generator.TZ_TO_WITH_TIME_ZONE = True
-
-
 @public
 class OracleCompiler(SQLGlotCompiler):
     __slots__ = ()
 
-    dialect = "oracle"
-    quoted = True
+    dialect = Oracle
     type_mapper = OracleType
     rewrites = (
         exclude_unsupported_window_frame_from_row_number,
