+++ ibis/backends/postgres/__init__.py
@@ -44,7 +44,6 @@ def _verify_source_line(func_name: str, line: str):
 
 class Backend(SQLGlotBackend):
     name = "postgres"
-    dialect = "postgres"
     compiler = PostgresCompiler()
     supports_python_udfs = True
