+++ ibis/backends/postgres/compiler.py
@@ -6,8 +6,6 @@ from functools import partial, reduce, singledispatchmethod
 import sqlglot as sg
 import sqlglot.expressions as sge
 from public import public
-from sqlglot.dialects import Postgres
-from sqlglot.dialects.dialect import rename_func
 
 import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
@@ -15,16 +13,8 @@ import ibis.expr.operations as ops
 import ibis.expr.rules as rlz
 from ibis.backends.base.sqlglot.compiler import NULL, STAR, SQLGlotCompiler, paren
 from ibis.backends.base.sqlglot.datatypes import PostgresType
+from ibis.backends.base.sqlglot.dialects import Postgres
 from ibis.backends.base.sqlglot.rewrites import rewrite_sample_as_filter
-
-Postgres.Generator.TRANSFORMS |= {
-    sge.Map: rename_func("hstore"),
-    sge.Split: rename_func("string_to_array"),
-    sge.RegexpSplit: rename_func("regexp_split_to_array"),
-    sge.DateFromParts: rename_func("make_date"),
-    sge.ArraySize: rename_func("cardinality"),
-    sge.Pow: rename_func("pow"),
-}
 
 
 class PostgresUDFNode(ops.Value):
@@ -35,10 +25,9 @@ class PostgresUDFNode(ops.Value):
 class PostgresCompiler(SQLGlotCompiler):
     __slots__ = ()
 
-    dialect = "postgres"
+    dialect = Postgres
     type_mapper = PostgresType
     rewrites = (rewrite_sample_as_filter, *SQLGlotCompiler.rewrites)
-    quoted = True
 
     NAN = sge.Literal.number("'NaN'::double precision")
     POS_INF = sge.Literal.number("'Inf'::double precision")
